mini_project/requirements.txt
flask==3.0.2
pymongo==4.6.1
python-dotenv==1.0.1
flask-pymongo==2.3.0
kafka-python==2.0.2
happybase==1.2.0
thrift==0.16.0 